requirements.txt
discord.py
python-dotenv
aiosqlite
aiohttp
pyyaml
matplotlib
numpy
seaborn
pandas
humanize
bar_chart_race
requests